requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
diffusers==0.16.1
discord.py==2.2.3
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
lit==16.0.5.post0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyre-extensions==0.0.23
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
xformers==0.0.16rc425
yarl==1.9.2
zipp==3.15.0
